+++ docs/example/src/bin/super-app.js
@@ -1,0 +1,4 @@
+#!/usr/bin/env node
+
+process.stdout.write("Hello world!\n");
+process.exit(0);