+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Added
+
+- CLI `workdir` parameter.
+
 ## [0.3.1] - 2020-05-08
 
 ### Added